Cleared up some compiler errors

+++ bankswap.c
@@ -11,7 +11,7 @@
 /* Process orders to swap bank A (port 06, 0x4000). */
 void bankswap_a_write(char data)
 {
-	mem_page_1 = pages[data];
+	mem_page_1 = pages[(int)data];
 	mem_page_loc_1 = data;
 	return;
 }
@@ -19,7 +19,7 @@ void bankswap_a_write(char data)
 /* Process orders to swap bank B (port 07, 0x8000). */
 void bankswap_b_write(char data)
 {
-	mem_page_2 = pages[data];
+	mem_page_2 = pages[(int)data];
 	mem_page_loc_2 = data;
 	return;
 }
